FEM: result package, move result tools module in new result package
